requirements.txt
emoji>=2.7.0
pytgasu>=0.2.1
pillow>=9.4.0
